edit profile

+++ src/utils/Api.js
@@ -89,34 +89,43 @@ class Api {
 
   }
 
-  deleteLikes(id) {
-    return fetch(`${this._baseUrl}/cards/${id}/likes`, {
-      method: 'DELETE',
-      headers: this._headers
-    })
-    .then(res => {
-      if (res.ok) {
-        return res.json();
-      }
-      // если ошибка, отклоняем промис
-    return Promise.reject(`Ошибка: ${res.status}`);
-    })
+  //deleteLikes(id) {
+    
 
-  }
+  //}
 
-  addLikes(id) {
-    return fetch(`${this._baseUrl}/cards/${id}/likes`, {
-      method: 'PUT',
-      headers: this._headers
-    })
-    .then(res => {
-      if (res.ok) {
-        return res.json();
-      }
-      // если ошибка, отклоняем промис
-    return Promise.reject(`Ошибка: ${res.status}`);
-    })
+  //addLikes(id) {
+    
 
+  //}
+
+  changeLikeCardStatus(id, isLiked) {
+    if(!isLiked) {
+      return fetch(`${this._baseUrl}/cards/${id}/likes`, {
+        method: 'DELETE',
+        headers: this._headers
+      })
+      .then(res => {
+        if (res.ok) {
+          return res.json();
+        }
+        // если ошибка, отклоняем промис
+      return Promise.reject(`Ошибка: ${res.status}`);
+      })
+      
+    } else {
+      return fetch(`${this._baseUrl}/cards/${id}/likes`, {
+        method: 'PUT',
+        headers: this._headers
+      })
+      .then(res => {
+        if (res.ok) {
+          return res.json();
+        }
+        // если ошибка, отклоняем промис
+      return Promise.reject(`Ошибка: ${res.status}`);
+      })
+    }
   }
 
   updateAvatar(avatar) {
